requirements.txt
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-dotenv==1.0.1
python-telegram-bot==20.7
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.2.0
pandas==2.2.0
bs4==0.0.2
beautifulsoup4==4.10.0
mysql-connector-python==8.3.0